roadie/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
geoalchemy2==0.14.2
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
shapely==2.0.2
geopy==2.4.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
websockets==12.0

